+++ tests/index_ops_test.py
@@ -71,16 +71,19 @@ def index_op8(t, idx, v=None):
     return t
 
 
-def index_harness(op, idxs, pass_value, v=None):
+def index_harness(op, idxs, is_index_put, v=None, is_mask=False):
     torch.manual_seed(42)
     t = torch.randn(2, 3, 4, 5)
-    idxs_tensors = [torch.tensor([i]) for i in idxs]
+    if not is_mask:
+        idxs_tensors = [torch.tensor([i]) for i in idxs]
+    else:
+        idxs_tensors = [idxs]
     model = helpers.ModelWithWeights(op, t.shape)
     # The LR should be large enough to guarantee weights change
     optim = torch.optim.AdamW(model.parameters(), lr=0.1)
     poptorch_model = poptorch.trainingModel(model, optimizer=optim)
 
-    if pass_value:
+    if is_index_put:
         if v is None:
             v = torch.zeros_like(op(t, *idxs_tensors))
         # Clone the tensor so that the original is unchanged by the in-place op
@@ -159,6 +162,20 @@ def test_index_put_broadcastable():
     v = torch.zeros(5)
     # For each row r in t[0, 0], r = [0, 0, 0, 0, 0]
     index_harness(index_op1, [[0]], True, v)
+
+
+@pytest.mark.parametrize("mask_size, dtype", [
+    (1, torch.bool),
+    (2, torch.uint8),
+    (3, torch.bool),
+    (4, torch.uint8),
+])
+def test_index_put_masked_fill(mask_size, dtype):
+    torch.manual_seed(42)
+    mask_shape = [2, 3, 4, 5][:mask_size]
+    mask = (torch.rand(mask_shape) > 0.5).type(dtype)
+    v = torch.tensor([0.])
+    index_harness(index_op0, mask, True, v=v, is_mask=True)
 
 
 @pytest.mark.parametrize("dim", range(-3, 3))